Added safety net for brownout detection and mitigation
SaturnRobot now includes an optionally overloadable
method called notifyBrownout. If there are 3 distinct brownouts detected
over a period of time, this method will be called. User code can then
use this method as a callback to cut back on power budget to mitigate
brownouts for the duration of the match